The parameter api should be uint32 for all callbacks

+++ include/pnet_api.h
@@ -651,7 +651,7 @@ typedef int (*pnet_read_ind)(
    pnet_t                  *net,
    void                    *arg,
    uint32_t                arep,
-   uint16_t                api,
+   uint32_t                api,
    uint16_t                slot,
    uint16_t                subslot,
    uint16_t                idx,
@@ -694,7 +694,7 @@ typedef int (*pnet_write_ind)(
    pnet_t                  *net,
    void                    *arg,
    uint32_t                arep,
-   uint16_t                api,
+   uint32_t                api,
    uint16_t                slot,
    uint16_t                subslot,
    uint16_t                idx,
@@ -730,7 +730,7 @@ typedef int (*pnet_write_ind)(
 typedef int (*pnet_exp_module_ind)(
    pnet_t                  *net,
    void                    *arg,
-   uint16_t                api,
+   uint32_t                api,
    uint16_t                slot,
    uint32_t                module_ident);
 
@@ -767,7 +767,7 @@ typedef int (*pnet_exp_module_ind)(
 typedef int (*pnet_exp_submodule_ind)(
    pnet_t                  *net,
    void                    *arg,
-   uint16_t                api,
+   uint32_t                api,
    uint16_t                slot,
    uint16_t                subslot,
    uint32_t                module_ident,
